Refactor POM & adjust package names
