+++ com.shemilikevin.app.tracker/pom.xml
@@ -15,39 +15,49 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
+
+    <mongo.java.driver.version>3.12.14</mongo.java.driver.version>
+    <logback-classic.version>1.2.13</logback-classic.version>
+    <junit.version>4.13.2</junit.version>
+    <assertj.core.version>3.27.3</assertj.core.version>
+    <mongo.java.server>1.39.0</mongo.java.server>
+
+    <pitest.maven.version>1.5.2</pitest.maven.version>
+    <jacoco.maven.plugin.version>0.8.12</jacoco.maven.plugin.version>
+
   </properties>
 
   <dependencies>
     <dependency>
       <groupId>org.mongodb</groupId>
       <artifactId>mongo-java-driver</artifactId>
-      <version>3.12.14</version>
+      <version>${mongo.java.driver.version}</version>
     </dependency>
 
     <dependency>
       <groupId>ch.qos.logback</groupId>
       <artifactId>logback-classic</artifactId>
-      <version>1.2.13</version>
+      <version>${logback-classic.version}</version>
     </dependency>
 
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
-      <version>4.13.2</version>
+      <version>${junit.version}</version>
       <scope>test</scope>
     </dependency>
 
     <dependency>
       <groupId>org.assertj</groupId>
       <artifactId>assertj-core</artifactId>
-      <version>3.27.3</version>
+      <version>${assertj.core.version}</version>
       <scope>test</scope>
     </dependency>
 
     <dependency>
       <groupId>de.bwaldvogel</groupId>
       <artifactId>mongo-java-server</artifactId>
-      <version>1.39.0</version>
+      <version>${mongo.java.server}</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
@@ -59,19 +69,19 @@
         <plugin>
           <groupId>org.pitest</groupId>
           <artifactId>pitest-maven</artifactId>
-          <version>1.5.2</version>
+          <version>${pitest.maven.version}</version>
           <configuration>
 
             <targetClasses>
-              <param>com.shemilikevin.app.tracker.repository.*</param>   <!-- PLACEHOLDER - CHANGEME -->
+              <param>com.shemilikevin.app.tracker.repository.*</param>
             </targetClasses>
 
             <targetTests>
-              <param>com.shemilikevin.app.tracker.repository.*</param>   <!-- PLACEHOLDER - CHANGEME -->
+              <param>com.shemilikevin.app.tracker.repository.*</param>
             </targetTests>
 
             <mutators>
-              <mutator>DEFAULTS</mutator>
+              <mutator>STRONGER</mutator>
             </mutators>
             <mutationThreshold>100</mutationThreshold>
           </configuration>
@@ -80,10 +90,11 @@
         <plugin>
           <groupId>org.jacoco</groupId>
           <artifactId>jacoco-maven-plugin</artifactId>
-          <version>0.8.12</version>
+          <version>${jacoco.maven.plugin.version}</version>
           <configuration>
             <excludes>
-              <exclude></exclude> <!-- PLACEHOLDER - CHANGEME -->
+              <exclude>**/model/*.*</exclude>
+              <exclude>**/App.*</exclude>
             </excludes>
           </configuration>
           <executions>
@@ -102,13 +113,13 @@
                 <rules>
                   <rule>
                     <element>PACKAGE</element>
-                    <limit>
+                    <limits>
                       <limit>
                         <counter>LINE</counter>
                         <value>COVEREDRATIO</value>
                         <minimum>1</minimum>
                       </limit>
-                    </limit>
+                    </limits>
                   </rule>
                 </rules>
               </configuration>
@@ -124,12 +135,6 @@
 
     <profile>
       <id>jacoco</id>
-      <activation>
-        <property>
-          <name>skipCoverage</name>
-          <value>!true</value>
-        </property>
-      </activation>
       <build>
         <plugins>
           <plugin>
