+++ com.shemilikevin.app.tracker/src/main/java/com/shemilikevin/app/tracker/app/swing/App.java
@@ -1,4 +1,4 @@
-package com.shemilikevin.app.tracker;
+package com.shemilikevin.app.tracker.app.swing;
 
 /**
  * Hello world!
