+++ com.shemilikevin.app.tracker/src/main/java/com/shemilikevin/app/tracker/model/Issue.java
@@ -1,4 +1,4 @@
-package com.shemilikevin.app.model;
+package com.shemilikevin.app.tracker.model;
 
 import java.util.Objects;
